chore: rename 'credentials' to 'credential'
