+++ client/src/app/modals/credentials/CredentialModal.css
@@ -8,15 +8,15 @@
  * except in compliance with the MIT License.
  */
 
-:local(.credentialsModal) {
-  .credentials-modal-loading {
+:local(.credentialModal) {
+  .credential-modal-loading {
     display: flex;
     align-items: center;
     justify-content: center;
     min-height: 240px;
   }
 
-  .credentials-modal-group {
+  .credential-modal-group {
     margin-top: 20px;
 
     h3 {
@@ -60,7 +60,7 @@
     }
   }
 
-  .credentials-modal-error {
+  .credential-modal-error {
     margin: 4px 0;
     font-size: 12px;
     line-height: 1.4;
